Initial Scraper (#38)
* Adding in an initial all recipes website scraper and setting up structure to be able to call out to python scripts.

* Adding scraper implementation.
This includes a scraper for a couple websites and integrating the
endpoint with them. Also updating the docker file while I'm here to be
able to run the python scraper logic. Updating the README to reflect the
changes. Adding tests to cover the logic.

* Adding untracked files
Putting one of the scrapers in, adding the requirements to ensure the
Dockerfile can install them, and adding a little password hash utility
file to get info for login endpoints and such.

* Fixing broken test
Was passing in the wrong user's password here.

* Test Updates
Adding an NPM install to the Dockerfile because it makes sense to do
that. Adding `cors` to the package.json file becau we need it.

* Fix test
Seems like the fact that we don't connect to the db is causing some
problems. Connecting in the test and will ignore if connection already
exists.

* Updating scraper logic and tests
Adding new tests to cover the scraper logic. This involves passing the
raw html downloaded from the websites to be "scraped" and then parsed
like in a real environment. Also using `json.dumps` now because double
quotes are required for NodeJS to parse the output and test correctly.

* Adding virutal env and pip install to workflow

* Remove create virtual env
Might not need it because we tear down the machine after each build.

* Trying to fix build with Python venv

* Starting virtual environment

* Trying to get virtual env working.

* Trying to get python working again.

* Removing the pkg-resources
Doesn't seem to be needed and it breaks the tests.

* Code clean up

Co-authored-by: michael <michael@lugia.lugia>